refactor: reorganize tab state management into separate modules
- Split monolithic tabState.ts into focused modules:
  - state/activationHistory.ts: Tab activation tracking
  - state/indexCache.ts: Tab index caching
  - state/sourceMap.ts: Tab source relationships
- Rename closeHandler.ts to tabClosing.ts for consistency
- Create state/ subdirectory for clear separation of concerns
- Maintain function-based approach instead of classes
